Extract dedicated services for GAQ and QC summaries
