+++ lib/server/controllers/qcFlag.controller.js
@@ -19,6 +19,8 @@ const { PaginationDto } = require('../../domain/dtos');
 const { ApiConfig } = require('../../config');
 const { countedItemsToHttpView } = require('../utilities/countedItemsToHttpView');
 const { qcFlagService } = require('../services/qualityControlFlag/QcFlagService.js');
+const { gaqService } = require('../services/qualityControlFlag/GaqService.js');
+const { qcFlagSummaryService } = require('../services/qualityControlFlag/QcFlagSummaryService.js');
 
 // eslint-disable-next-line valid-jsdoc
 /**
@@ -324,7 +326,7 @@ const getQcFlagsSummaryHandler = async (req, res) => {
                 mcReproducibleAsNotBad = false,
             } = validatedDTO.query;
 
-            const data = await qcFlagService.getQcFlagsSummary({ dataPassId, simulationPassId, lhcPeriodId }, { mcReproducibleAsNotBad });
+            const data = await qcFlagSummaryService.getQcFlagsSummary({ dataPassId, simulationPassId, lhcPeriodId }, { mcReproducibleAsNotBad });
             res.json({ data });
         } catch (error) {
             updateExpressResponseFromNativeError(res, error);
@@ -349,7 +351,7 @@ const getGaqQcFlagsHandler = async (request, response) => {
         try {
             const { dataPassId, runNumber } = validatedDTO.query;
 
-            const data = await qcFlagService.getGaqFlags(dataPassId, runNumber);
+            const data = await gaqService.getFlagsForDataPassAndRun(dataPassId, runNumber);
             response.json({ data });
         } catch (error) {
             updateExpressResponseFromNativeError(response, error);
@@ -374,7 +376,7 @@ const getGaqSummaryHandler = async (request, response) => {
         try {
             const { dataPassId, mcReproducibleAsNotBad = false } = validatedDTO.query;
 
-            const data = await qcFlagService.getGaqSummary(dataPassId, { mcReproducibleAsNotBad });
+            const data = await gaqService.getSummary(dataPassId, { mcReproducibleAsNotBad });
             response.json({ data });
         } catch (error) {
             updateExpressResponseFromNativeError(response, error);
