+++ lib/server/services/qualityControlFlag/QcFlagService.js
@@ -64,20 +64,6 @@ const validateUserDetectorAccess = (userRoles, detectorName) => {
 };
 
 /**
- * @typedef RunDetectorQcSummary
- * @property {number} badEffectiveRunCoverage - fraction of run's data, marked explicitly with bad QC flag
- * @property {number} explicitlyNotBadEffectiveRunCoverage - fraction of run's data, marked explicitly with good QC flag
- * @property {number} missingVerificationsCount - number of not verified QC flags which are not discarded
- * @property {boolean} mcReproducible - states whether some Limited Acceptance MC Reproducible flag was assigned
- */
-const QC_SUMMARY_PROPERTIES = {
-    badEffectiveRunCoverage: 'badEffectiveRunCoverage',
-    explicitlyNotBadEffectiveRunCoverage: 'explicitlyNotBadEffectiveRunCoverage',
-    missingVerificationsCount: 'missingVerificationsCount',
-    mcReproducible: 'mcReproducible',
-};
-
-/**
  * @typedef RunQcSummary
  * @type {Object<number, RunDetectorQcSummary>} dplDetectorID to RunDetectorQcSummary mappings
  */
@@ -104,6 +90,14 @@ const QC_SUMMARY_PROPERTIES = {
  */
 
 /**
+ * @typedef RunDetectorQcSummary
+ * @property {number} badEffectiveRunCoverage - fraction of run's data, marked explicitly with bad QC flag
+ * @property {number} explicitlyNotBadEffectiveRunCoverage - fraction of run's data, marked explicitly with good QC flag
+ * @property {number} missingVerificationsCount - number of not verified QC flags which are not discarded
+ * @property {boolean} mcReproducible - states whether some Limited Acceptance MC Reproducible flag was assigned
+ */
+
+/**
  * @typedef GaqSummary aggregated global quality summaries for given data pass
  * @type {Object<number, RunGaqSummary>} runNumber to RunGaqSummary mapping
  */
@@ -120,7 +114,7 @@ class QcFlagService {
     }
 
     /**
-     * Find an Quality Control Flag by its id
+     * Find a Quality Control Flag by its id
      * @param {number} id identifier of Quality Control Flag
      * @return {QcFlag} a Quality Control Flag
      */
@@ -137,7 +131,7 @@ class QcFlagService {
     }
 
     /**
-     * Find an Quality Control Flag by its id
+     * Find a Quality Control Flag by its id
      * @param {number} id id of Quality Control Flag
      * @throws {NotFoundError} in case there is no Quality Control Flag with given id
      * @return {Promise<QcFlag>} a Quality Control Flag
@@ -148,204 +142,6 @@ class QcFlagService {
             throw new NotFoundError(`Quality Control Flag with this id (${id}) could not be found`);
         }
         return qcFlag;
-    }
-
-    /**
-     * Validate QC flag timestamps
-     * If null timestamp was provided, given timestamp is replaced by run's startTime or endTime
-     * @param {Partial<Period>} timestamps QC flag timestamps
-     * @param {Run} targetRun run which for QC flag is to be set
-     * @return {{from: (number|null), to: (number|null)}} prepared timestamps
-     * @throws {BadParameterError}
-     */
-    _prepareQcFlagPeriod(timestamps, targetRun) {
-        const { timeTrgStart, timeO2Start, firstTfTimestamp, timeTrgEnd, timeO2End, lastTfTimestamp } = targetRun;
-
-        let lowerBound = timeTrgStart?.getTime() ?? timeO2Start?.getTime() ?? null;
-        if (firstTfTimestamp) {
-            lowerBound = lowerBound ? Math.min(firstTfTimestamp.getTime(), lowerBound) : firstTfTimestamp.getTime();
-        }
-
-        let upperBound = timeTrgEnd?.getTime() ?? timeO2End?.getTime() ?? null;
-        if (lastTfTimestamp) {
-            upperBound = upperBound ? Math.max(lastTfTimestamp.getTime(), upperBound) : lastTfTimestamp.getTime();
-        }
-
-        const from = timestamps.from ?? lowerBound ?? null;
-        const to = timestamps.to ?? upperBound ?? null;
-
-        if (from && to && from >= to) {
-            throw new BadParameterError('Parameter "to" timestamp must be greater than "from" timestamp');
-        }
-        const isFromOutOfRange = from && (lowerBound && from < lowerBound || upperBound && upperBound <= from);
-        const isToOutOfRange = to && (lowerBound && to <= lowerBound || upperBound && upperBound < to);
-        if (isFromOutOfRange || isToOutOfRange) {
-            throw new BadParameterError(`Given QC flag period (${from}, ${to}) is out of run (${lowerBound}, ${upperBound}) period`);
-        }
-
-        return { from, to };
-    }
-
-    /**
-     * Remove a time segment from a list of QC flags effective periods
-     *
-     * @param {Period} intersectingPeriod time segment that should be removed from given periods
-     * @param {SequelizeQcFlagEffectivePeriod} periods periods to be updated or discarded
-     * @return {Promise<void>} resolve once all periods are updated
-     */
-    async _removeEffectivePeriodsAndPeriodIntersection({ from: newerPeriodFrom, to: newerPeriodTo }, periods) {
-        for (const effectivePeriod of periods) {
-            const { id: effectivePeriodId, from: effectiveFrom, to: effectiveTo } = effectivePeriod;
-
-            const effectiveToIsLesserOrEqNewerPeriodTo = newerPeriodTo === null || effectiveTo !== null && effectiveTo <= newerPeriodTo;
-
-            if (newerPeriodFrom <= effectiveFrom
-                && effectiveToIsLesserOrEqNewerPeriodTo) { // Old flag is fully covered by new one
-                await QcFlagEffectivePeriodRepository.removeOne({ where: { id: effectivePeriodId } });
-            } else if (effectiveFrom < newerPeriodFrom && !effectiveToIsLesserOrEqNewerPeriodTo) {
-                // New flag's period is included in the old one's period.
-                await QcFlagEffectivePeriodRepository.update(effectivePeriod, { to: newerPeriodFrom });
-                await QcFlagEffectivePeriodRepository.insert({
-                    flagId: effectivePeriod.flagId,
-                    from: newerPeriodTo,
-                    to: effectiveTo,
-                });
-            } else if (effectiveFrom < newerPeriodFrom) {
-                await QcFlagEffectivePeriodRepository.update(effectivePeriod, { to: newerPeriodFrom });
-            } else if (!effectiveToIsLesserOrEqNewerPeriodTo) {
-                await QcFlagEffectivePeriodRepository.update(effectivePeriod, { from: newerPeriodTo });
-            } else {
-                throw new Error('Incorrect state');
-            }
-        }
-    }
-
-    /**
-     * Get QC summary for given data/simulation pass or synchronous QC flags for given LHC period
-     *
-     * @param {scope} scope of the QC flag
-     * @param {number} [scope.dataPassId] data pass id - exclusive with other options
-     * @param {number} [scope.simulationPassId] simulation pass id - exclusive with other options
-     * @param {number} [scope.lhcPeriodId] id of LHC Period - exclusive with other options
-     * @param {object} [options] additional options
-     * @param {boolean} [options.mcReproducibleAsNotBad = false] if set to true, `Limited Acceptance MC Reproducible` flag type is treated as
-     *     good one
-     * @return {Promise<QcSummary[]>} summary
-     */
-    async getQcFlagsSummary({ dataPassId, simulationPassId, lhcPeriodId }, { mcReproducibleAsNotBad = false } = {}) {
-        if (Boolean(dataPassId) + Boolean(simulationPassId) + Boolean(lhcPeriodId) > 1) {
-            throw new BadParameterError('`dataPassId`, `simulationPassId` and `lhcPeriodId` are exclusive options');
-        }
-
-        const queryBuilder = dataSource.createQueryBuilder()
-            .where('deleted').is(false);
-
-        if (dataPassId) {
-            queryBuilder
-                .whereAssociation('dataPasses', 'id').is(dataPassId)
-                .include({ association: 'run', attributes: [] });
-        } else if (simulationPassId) {
-            queryBuilder
-                .whereAssociation('simulationPasses', 'id').is(simulationPassId)
-                .include({ association: 'run', attributes: [] });
-        } else {
-            queryBuilder.include({ association: 'dataPasses', required: false })
-                .include({ association: 'simulationPasses', required: false })
-                .where('$dataPasses.id$').is(null)
-                .where('$simulationPasses.id$').is(null)
-                .include({ association: 'run', attributes: [], where: { lhcPeriodId } });
-        }
-
-        queryBuilder
-            .include({ association: 'effectivePeriods', attributes: [], required: true })
-            .include({ association: 'flagType', attributes: [] })
-            .set('attributes', (sequelize) => [
-                'runNumber',
-                'detectorId',
-                [sequelize.literal(`IF(\`flagType\`.monte_carlo_reproducible AND ${mcReproducibleAsNotBad}, false, \`flagType\`.bad)`), 'bad'],
-
-                [
-                    sequelize.literal(`
-                    IF(
-                        run.time_start IS NULL OR run.time_end IS NULL,
-                        IF(
-                            effectivePeriods.\`from\` IS NULL AND effectivePeriods.\`to\` IS NULL,
-                            1,
-                            null
-                        ),
-                        SUM(
-                            UNIX_TIMESTAMP(COALESCE(effectivePeriods.\`to\`, run.time_end))
-                          - UNIX_TIMESTAMP(COALESCE(effectivePeriods.\`from\`, run.time_start))
-                        ) / (
-                            UNIX_TIMESTAMP(run.time_end) - UNIX_TIMESTAMP(run.time_start)
-                        )
-                    )
-                    `),
-                    'effectiveRunCoverage',
-                ],
-                [
-                    sequelize.literal('GROUP_CONCAT( DISTINCT `QcFlag`.id )'),
-                    'flagIds',
-                ],
-                [
-                    sequelize.literal('SUM( `flagType`.monte_carlo_reproducible ) > 0'),
-                    'mcReproducible',
-                ],
-            ])
-            .groupBy('runNumber')
-            .groupBy('detectorId')
-            .groupBy((sequelize) => sequelize.literal(`
-                IF(\`flagType\`.monte_carlo_reproducible AND ${mcReproducibleAsNotBad}, false, \`flagType\`.bad)
-            `));
-
-        const runDetectorSummaryList = (await QcFlagRepository.findAll(queryBuilder))
-            .map((summaryDb) =>
-                ({
-                    runNumber: summaryDb.runNumber,
-                    detectorId: summaryDb.detectorId,
-                    effectiveRunCoverage: parseFloat(summaryDb.get('effectiveRunCoverage'), 10) || null,
-                    bad: Boolean(summaryDb.get('bad')),
-                    flagIds: (summaryDb.get('flagIds')?.split(',') ?? []).map((id) => parseInt(id, 10)),
-                    mcReproducible: Boolean(summaryDb.get('mcReproducible')),
-                }));
-
-        const allFlagsIds = new Set(runDetectorSummaryList.flatMap(({ flagIds }) => flagIds));
-        const notVerifiedFlagsIds = new Set((await QcFlagRepository.findAll({
-            attributes: ['id'],
-            include: [{ association: 'verifications', required: false, attributes: [] }],
-            where: {
-                id: { [Op.in]: [...allFlagsIds] },
-                '$verifications.id$': { [Op.is]: null },
-            },
-        })).map(({ id }) => id));
-
-        const summary = {};
-
-        // Fold list of summaries into nested object
-        for (const runDetectorSummaryForFlagTypesClass of runDetectorSummaryList) {
-            const {
-                runNumber,
-                detectorId,
-                flagIds,
-            } = runDetectorSummaryForFlagTypesClass;
-            const missingVerificationsCount = flagIds.filter((id) => notVerifiedFlagsIds.has(id)).length;
-
-            if (!summary[runNumber]) {
-                summary[runNumber] = {};
-            }
-            if (!summary[runNumber][detectorId]) {
-                summary[runNumber][detectorId] = { [QC_SUMMARY_PROPERTIES.mcReproducible]: false };
-            }
-
-            const runDetectorSummary = summary[runNumber][detectorId];
-
-            runDetectorSummary[QC_SUMMARY_PROPERTIES.missingVerificationsCount] =
-                (runDetectorSummary[QC_SUMMARY_PROPERTIES.missingVerificationsCount] ?? 0) + missingVerificationsCount;
-
-            this._mergeIntoSummaryUnit(runDetectorSummary, runDetectorSummaryForFlagTypesClass);
-        }
-
-        return summary;
     }
 
     /**
@@ -659,120 +455,6 @@ class QcFlagService {
     }
 
     /**
-     * Find QC flags in GAQ effective periods for given data pass and run
-     *
-     * @param {number} dataPassId id od data pass
-     * @param {number} runNumber run number
-     * @return {Promise<GaqFlags[]>} promise of aggregated QC flags
-     */
-    async getGaqFlags(dataPassId, runNumber) {
-        const gaqPeriods = await QcFlagRepository.findGaqPeriods(dataPassId, runNumber);
-        const qcFlags = (await QcFlagRepository.findAll({
-            where: { id: { [Op.in]: gaqPeriods.flatMap(({ contributingFlagIds }) => contributingFlagIds) } },
-            include: [
-                { association: 'flagType' },
-                { association: 'createdBy' },
-                { association: 'verifications', include: [{ association: 'createdBy' }] },
-            ],
-        })).map(qcFlagAdapter.toEntity);
-
-        const idToFlag = Object.fromEntries(qcFlags.map((flag) => [flag.id, flag]));
-
-        return gaqPeriods.map(({
-            contributingFlagIds,
-            from,
-            to,
-        }) => ({
-            from,
-            to,
-            contributingFlags: contributingFlagIds.map((id) => idToFlag[id]),
-        }));
-    }
-
-    /**
-     * Get GAQ summary
-     *
-     * @param {number} dataPassId id of data pass id
-     * @param {object} [options] additional options
-     * @param {boolean} [options.mcReproducibleAsNotBad = false] if set to true,
-     * `Limited Acceptance MC Reproducible` flag type is treated as good one
-     * @return {Promise<GaqSummary[]>} Resolves with the GAQ Summary
-     */
-    async getGaqSummary(dataPassId, { mcReproducibleAsNotBad = false } = {}) {
-        await getOneDataPassOrFail({ id: dataPassId });
-        const runGaqSubSummaries = await QcFlagRepository.getRunGaqSubSummaries(dataPassId, { mcReproducibleAsNotBad });
-
-        const summary = {};
-        const flagsAndVerifications = {};
-
-        // Fold list of subSummaries into one summary
-        for (const subSummary of runGaqSubSummaries) {
-            const {
-                runNumber,
-                flagsIds,
-                verifiedFlagsIds,
-            } = subSummary;
-
-            if (!summary[runNumber]) {
-                summary[runNumber] = { [QC_SUMMARY_PROPERTIES.mcReproducible]: false };
-            }
-            if (!flagsAndVerifications[runNumber]) {
-                flagsAndVerifications[runNumber] = {};
-            }
-
-            const runSummary = summary[runNumber];
-
-            const distinctRunFlagsIds = flagsAndVerifications[runNumber]?.distinctFlagsIds ?? [];
-            const distinctRunVerifiedFlagsIds = flagsAndVerifications[runNumber]?.distinctVerifiedFlagsIds ?? [];
-
-            flagsAndVerifications[runNumber] = {
-                distinctFlagsIds: new Set([...distinctRunFlagsIds, ...flagsIds]),
-                distinctVerifiedFlagsIds: new Set([...distinctRunVerifiedFlagsIds, ...verifiedFlagsIds]),
-            };
-
-            this._mergeIntoSummaryUnit(runSummary, subSummary);
-        }
-
-        for (const [runNumber, { distinctFlagsIds, distinctVerifiedFlagsIds }] of Object.entries(flagsAndVerifications)) {
-            summary[runNumber][QC_SUMMARY_PROPERTIES.missingVerificationsCount] = distinctFlagsIds.size - distinctVerifiedFlagsIds.size;
-        }
-
-        return summary;
-    }
-
-    /**
-     * Update RunDetectorQcSummary or RunGaqSummary with new information
-     *
-     * @param {RunDetectorQcSummary|RunGaqSummary} summaryUnit RunDetectorQcSummary or RunGaqSummary
-     * @param {{ bad: boolean, effectiveRunCoverage: number, mcReproducible: boolean}} partialSummaryUnit new properties
-     * to be applied to the summary object
-     * @return {void}
-     */
-    _mergeIntoSummaryUnit(summaryUnit, partialSummaryUnit) {
-        const {
-            bad,
-            effectiveRunCoverage,
-            mcReproducible,
-        } = partialSummaryUnit;
-
-        if (bad) {
-            summaryUnit[QC_SUMMARY_PROPERTIES.badEffectiveRunCoverage] = effectiveRunCoverage;
-            summaryUnit[QC_SUMMARY_PROPERTIES.mcReproducible] =
-                mcReproducible || summaryUnit[QC_SUMMARY_PROPERTIES.mcReproducible];
-        } else {
-            summaryUnit[QC_SUMMARY_PROPERTIES.explicitlyNotBadEffectiveRunCoverage] = effectiveRunCoverage;
-            summaryUnit[QC_SUMMARY_PROPERTIES.mcReproducible] =
-                mcReproducible || summaryUnit[QC_SUMMARY_PROPERTIES.mcReproducible];
-        }
-        if (summaryUnit[QC_SUMMARY_PROPERTIES.badEffectiveRunCoverage] === undefined) {
-            summaryUnit[QC_SUMMARY_PROPERTIES.badEffectiveRunCoverage] = 0;
-        }
-        if (summaryUnit[QC_SUMMARY_PROPERTIES.explicitlyNotBadEffectiveRunCoverage] === undefined) {
-            summaryUnit[QC_SUMMARY_PROPERTIES.explicitlyNotBadEffectiveRunCoverage] = 0;
-        }
-    }
-
-    /**
      * Return a paginated list of QC flags related to a given simulation pass, run and dpl detector
      *
      * @param {object} criteria the QC flag criteria
@@ -881,6 +563,76 @@ class QcFlagService {
             .include({ association: 'createdBy' })
             .include({ association: 'verifications', include: [{ association: 'createdBy' }] })
             .orderBy('createdAt', 'DESC', 'verifications');
+    }
+
+    /**
+     * Validate QC flag timestamps
+     * If null timestamp was provided, given timestamp is replaced by run's startTime or endTime
+     * @param {Partial<Period>} timestamps QC flag timestamps
+     * @param {Run} targetRun run which for QC flag is to be set
+     * @return {{from: (number|null), to: (number|null)}} prepared timestamps
+     * @throws {BadParameterError}
+     */
+    _prepareQcFlagPeriod(timestamps, targetRun) {
+        const { timeTrgStart, timeO2Start, firstTfTimestamp, timeTrgEnd, timeO2End, lastTfTimestamp } = targetRun;
+
+        let lowerBound = timeTrgStart?.getTime() ?? timeO2Start?.getTime() ?? null;
+        if (firstTfTimestamp) {
+            lowerBound = lowerBound ? Math.min(firstTfTimestamp.getTime(), lowerBound) : firstTfTimestamp.getTime();
+        }
+
+        let upperBound = timeTrgEnd?.getTime() ?? timeO2End?.getTime() ?? null;
+        if (lastTfTimestamp) {
+            upperBound = upperBound ? Math.max(lastTfTimestamp.getTime(), upperBound) : lastTfTimestamp.getTime();
+        }
+
+        const from = timestamps.from ?? lowerBound ?? null;
+        const to = timestamps.to ?? upperBound ?? null;
+
+        if (from && to && from >= to) {
+            throw new BadParameterError('Parameter "to" timestamp must be greater than "from" timestamp');
+        }
+        const isFromOutOfRange = from && (lowerBound && from < lowerBound || upperBound && upperBound <= from);
+        const isToOutOfRange = to && (lowerBound && to <= lowerBound || upperBound && upperBound < to);
+        if (isFromOutOfRange || isToOutOfRange) {
+            throw new BadParameterError(`Given QC flag period (${from}, ${to}) is out of run (${lowerBound}, ${upperBound}) period`);
+        }
+
+        return { from, to };
+    }
+
+    /**
+     * Remove a time segment from a list of QC flags effective periods
+     *
+     * @param {Period} intersectingPeriod time segment that should be removed from given periods
+     * @param {SequelizeQcFlagEffectivePeriod} periods periods to be updated or discarded
+     * @return {Promise<void>} resolve once all periods are updated
+     */
+    async _removeEffectivePeriodsAndPeriodIntersection({ from: newerPeriodFrom, to: newerPeriodTo }, periods) {
+        for (const effectivePeriod of periods) {
+            const { id: effectivePeriodId, from: effectiveFrom, to: effectiveTo } = effectivePeriod;
+
+            const effectiveToIsLesserOrEqNewerPeriodTo = newerPeriodTo === null || effectiveTo !== null && effectiveTo <= newerPeriodTo;
+
+            if (newerPeriodFrom <= effectiveFrom
+                && effectiveToIsLesserOrEqNewerPeriodTo) { // Old flag is fully covered by new one
+                await QcFlagEffectivePeriodRepository.removeOne({ where: { id: effectivePeriodId } });
+            } else if (effectiveFrom < newerPeriodFrom && !effectiveToIsLesserOrEqNewerPeriodTo) {
+                // New flag's period is included in the old one's period.
+                await QcFlagEffectivePeriodRepository.update(effectivePeriod, { to: newerPeriodFrom });
+                await QcFlagEffectivePeriodRepository.insert({
+                    flagId: effectivePeriod.flagId,
+                    from: newerPeriodTo,
+                    to: effectiveTo,
+                });
+            } else if (effectiveFrom < newerPeriodFrom) {
+                await QcFlagEffectivePeriodRepository.update(effectivePeriod, { to: newerPeriodFrom });
+            } else if (!effectiveToIsLesserOrEqNewerPeriodTo) {
+                await QcFlagEffectivePeriodRepository.update(effectivePeriod, { from: newerPeriodTo });
+            } else {
+                throw new Error('Incorrect state');
+            }
+        }
     }
 
     /**
