+++ lib/usecases/run/GetAllRunsUseCase.js
@@ -20,7 +20,7 @@ const sequelize = require('sequelize');
 const { EorReasonRepository } = require('../../database/repositories');
 const { PhysicalConstant } = require('../../domain/enums/PhysicalConstant');
 const { BadParameterError } = require('../../server/errors/BadParameterError');
-const { qcFlagService } = require('../../server/services/qualityControlFlag/QcFlagService');
+const { gaqService } = require('../../server/services/qualityControlFlag/GaqService.js');
 
 /**
  * GetAllRunsUseCase
@@ -343,7 +343,7 @@ class GetAllRunsUseCase {
                 }
                 const { mcReproducibleAsNotBad = false } = gaq;
                 const [dataPassId] = dataPassIds;
-                const gaqSummary = await qcFlagService.getGaqSummary(dataPassId, { mcReproducibleAsNotBad });
+                const gaqSummary = await gaqService.getSummary(dataPassId, { mcReproducibleAsNotBad });
 
                 for (const [operator, value] of Object.entries(gaq.notBadFraction)) {
                     const runNumbers = Object.entries(gaqSummary)
